# Conflicts:
#	game/client/headers/MapUi.h
#	game/client/sources/CreateBarracks.cpp
#	game/client/sources/MapUi.cpp
#	game/server/sources/map/ServerMap.cpp
#	game/server/sources/model/ServerProtocol.cpp

+++ game/common/headers/Entity.h
@@ -7,7 +7,7 @@
 class Entity {  // Va terminar siendo una clase abstracta
 protected:
     int id;
-    double HP;
+    unsigned int HP;
     int type;
     coordenada_t coord;
 
@@ -15,7 +15,7 @@ public:
     int getType() const;
     int getId() const;
     coordenada_t getPosition();
-    Entity(int id, double HP, int type, coordenada_t coord);
+    Entity(int id, unsigned int HP, int type, coordenada_t coord);
 
     Entity();
